\name{aldh2}
\docType{data}
\alias{aldh2}
\title{ALDH2 markers and Alcoholism}
\description{This data set contains eight ALDH2 markers and Japanese alcohlic
patients (y=1) and controls (y=0). There are genotypes for 8 loci, with a prefix 
name (e.g., "EXON12") and a suffix for each of two alleles (".a1" and ".a2").

The eight markers loci follows the following map (base pairs)

\tabular{lr}{
D12S2070   \tab (> 450 000),\cr
D12S839    \tab (> 450 000),\cr
D12S821    \tab (\eqn{\sim}{~} 400 000),\cr
D12S1344   \tab (   83 853),\cr
EXON12     \tab (    0),\cr
EXON1      \tab (   37 335),\cr
D12S2263   \tab (   38 927),\cr
D12S1341   \tab (> 450 000)
}
}

\usage{data(aldh2)}
\format{A data frame}
\source{Prof Ian Craig of Oxford and SGDP Centre, KCL}
  
\references{
Koch HG, McClay J, Loh E-W, Higuchi S, Zhao J-H, Sham P, Ball D, et 
al (2000) Allele association studies with SSR and SNP markers 
at known physical distances within a 1 Mb region embracing the 
ALDH2 locus in the Japanese, demonstrates linkage disequilibrium 
extending up to 400 kb. Hum. Mol. Genet. 9:2993-2999
}
\keyword{datasets}
