\name{fsnps}
\docType{data}
\alias{fsnps}
\title{A case-control data involving four SNPs with missing genotype}
\description{
This is a simulated data of four SNPs with their alleles coded in characters.
The variable y contains phenotypes (1=case, 0=control).
}
\usage{data(fsnps)}
\format{A data frame}
\source{Dr Sebastien Lissarrague of Genset}
  
\keyword{datasets}
