\name{hla}
\docType{data}
\alias{hla}
\title{The HLA data}
\description{This data set contains HLA markers DRB, DQA, DQB and phenotypes
of 271 Schizophrenia patients (y=1) and controls (y=0). Genotypes for 3 HLA
loci have prefixes name (e.g., "DQB") and a suffix for each of two alleles 
(".a1" and ".a2").
}
\usage{data(hla)}
\format{A data frame containing 271 rows and 8 columns}
\source{Dr Padraig Wright of Pfizer}
  
\keyword{datasets}
