\name{hr1420}
\docType{data}
\alias{hr1420}
\title{An example data for Manhattan plot with annotation (mhtplot2)}
\description{
This example contains p values for a list of SNPs wtih information 
on chromosome, position and gene symnol.

In the reference below, seven established SNPs are in light blue, 14 
new SNPs in dark blue and those failed to replicate in red. The paper 
size is set to 189 width x 189/2 height (mm) and 1200 dpi resolution.
The font is Verdana. See \code{\link[gap]{mhtplot2}} for all the details.
}
\usage{data(hr1420)}
\format{A data frame}
\examples{
head(hr1420)
}
\source{Dr Marcel den Hoed at the MRC Epidemiology Unit}
\references{
de Hoed M et al. (2013) Heart rate-associated loci and their effects on cardiac conduction and rhythm
disorders. Nature Genetics 45(6):621-31, doi: 10.1038/ng.2610.
}
\keyword{datasets}
