\name{mao}
\docType{data}
\alias{mao}
\title{A study of Parkinson's disease and MAO gene}
\description{
The markers are both with actual allele sizes and allele numbers. The
dataset is distributed with the GENECOUNTING version 2.0 illustrating
gene counting method involving chromosome X. A total of 183 patients and 157
controls (150 males, 190 females) were available, together with five markers
in MAOA (monoamine oxidase A) region with alleles 12, 9, 6, 5, 3, and the
first three markers were genotyped in all individuals while the fourth and
fifth were genotyped for 294 and 304 individuals.

}
\usage{data(mao)}
\format{A data frame}
\source{Dr Helen Latsoudis of Institute of Psychiatry, KCL}
\references{
Zhao JH (2004). 2LD, GENECOUNTING and HAP: computer programs for linkage
disequilibrium analysis. Bioinformatics 20:1325-1326
}
\keyword{datasets}
