\name{mhtdata}
\docType{data}
\alias{mhtdata}
\title{An example data for Manhattan plot with annotation (mhtplot)}
\description{
This example contains p values for a list of SNPs whose information 
regarding chromosome, position and reference seqeuence as with gene
annotation is obtained separately.}
\usage{data(mhtdata)}
\format{A data frame}
\examples{
head(mhtdata)
}
\source{Dr Tuomas Kilpelainen at the MRC Epidemiology Unit}
\references{
Kilpelainen1 TO, et al. (2011) Genetic variation near IRS1 associates with reduced adiposity 
and an impaired metabolic profile. Nature Genetics 43(8):753-60, doi: 10.1038/ng.866.
} 
\keyword{datasets}
