\name{OPG}
\docType{data}
\alias{OPG}
\alias{OPGtbl}
\alias{OPGall}
\alias{OPGrsid}
\title{An example data for forest plot using METAL output (METAL_forestplot)}
\description{
This example contains METAL outputs (tbl) as with association statistics from contributing studies (all).
It is appropriate to use chr:pos_A1_A2 (A1<=A2) (SNPID) rather than reference id (rsid) due to its 
variability -- therefore a SNPID-rsid mapping file (rsid) is also provided.
}
\usage{data(OPG)}
\format{Three data frames}
\examples{
head(OPGtbl)
head(OPGall)
head(OPGrsid)
}
\source{SCALLOP consortium}
\references{
pending to give.
}
\keyword{datasets}
