\name{w4}
\docType{data}
\alias{w4}
\title{Results from a GWAS on Chickens}
\description{
This example contains p values for a list of SNPs wtih information 
on chromosome and positions.
}
\usage{data(w4)}
\format{A data frame}
\examples{
head(w4)
}
\source{Guo Jun <guojun.yz@gmail.com>}
\keyword{datasets}
