\name{OPG}
\docType{data}
\alias{OPG}
\alias{tbl}
\alias{all}
\alias{rsid}
\title{An example data for forest plot using METAL output (METAL_forestplot)}
\description{
This example contains cojo outputs as with association statistics from contributing studies.

See \code{\link[gap]{METAL_forestplot}} for its use.
}
\usage{data(OPG)}
\format{Three data frames}
\examples{
head(tbl)
head(all)
head(rsid)
}
\source{SCALLOP consortium}
\references{
pending to give.
}
\keyword{datasets}
